+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v0.7.3
 * Python plotting script for 1-D flows now allows for multiple files + plot labels
+* `physical_props.jl', `surface_props.jl`, `flux_props.jl` files moved to new `src/properties` directory
+* Added computation of mean collision frequency and mean free path for single-species VHS gases: `mean_free_path`,
+`mean_collision_frequency`.
 
 ## v0.7.2
 * Added `w_threshold` keyword argument to NNLS merging to discard particles with very small weights